Move githubapp package
