+++ cmd/frontend/internal/batches/resolvers/BUILD.bazel
@@ -44,7 +44,7 @@ go_library(
         "//cmd/frontend/graphqlbackend",
         "//cmd/frontend/graphqlbackend/externallink",
         "//cmd/frontend/graphqlbackend/graphqlutil",
-        "//cmd/frontend/internal/auth/githubappauth",
+        "//cmd/frontend/internal/githubapp",
         "//internal/actor",
         "//internal/api",
         "//internal/auth",
@@ -132,8 +132,8 @@ go_test(
         "//cmd/frontend/globals",
         "//cmd/frontend/graphqlbackend",
         "//cmd/frontend/graphqlbackend/externallink",
-        "//cmd/frontend/internal/auth/githubappauth",
         "//cmd/frontend/internal/batches/resolvers/apitest",
+        "//cmd/frontend/internal/githubapp",
         "//internal/actor",
         "//internal/api",
         "//internal/batches/graphql",
