+++ cmd/frontend/internal/githubapp/BUILD.bazel
@@ -71,3 +71,70 @@ go_test(
         "@com_github_stretchr_testify//require",
     ],
 )
+
+go_library(
+    name = "githubapp",
+    srcs = [
+        "httpapi.go",
+        "init.go",
+        "resolver.go",
+    ],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/frontend/internal/githubapp",
+    visibility = ["//cmd/frontend:__subpackages__"],
+    deps = [
+        "//cmd/frontend/backend",
+        "//cmd/frontend/enterprise",
+        "//cmd/frontend/graphqlbackend",
+        "//cmd/frontend/internal/repos/webhooks/resolvers",
+        "//internal/auth",
+        "//internal/codeintel",
+        "//internal/conf/conftypes",
+        "//internal/database",
+        "//internal/encryption",
+        "//internal/encryption/keyring",
+        "//internal/extsvc",
+        "//internal/extsvc/github",
+        "//internal/github_apps/auth",
+        "//internal/github_apps/types",
+        "//internal/gqlutil",
+        "//internal/httpcli",
+        "//internal/observation",
+        "//internal/rcache",
+        "//internal/types",
+        "//lib/errors",
+        "//schema",
+        "@com_github_google_uuid//:uuid",
+        "@com_github_gorilla_mux//:mux",
+        "@com_github_graph_gophers_graphql_go//:graphql-go",
+        "@com_github_graph_gophers_graphql_go//relay",
+        "@com_github_sourcegraph_log//:log",
+    ],
+)
+
+go_test(
+    name = "githubapp_test",
+    srcs = [
+        "httpapi_test.go",
+        "resolver_test.go",
+    ],
+    embed = [":githubapp"],
+    deps = [
+        "//cmd/frontend/graphqlbackend",
+        "//internal/actor",
+        "//internal/database/dbmocks",
+        "//internal/encryption",
+        "//internal/github_apps/store",
+        "//internal/github_apps/types",
+        "//internal/gitserver",
+        "//internal/httpcli",
+        "//internal/rcache",
+        "//internal/types",
+        "//lib/errors",
+        "@com_github_google_uuid//:uuid",
+        "@com_github_gorilla_mux//:mux",
+        "@com_github_graph_gophers_graphql_go//errors",
+        "@com_github_sourcegraph_log//logtest",
+        "@com_github_stretchr_testify//assert",
+        "@com_github_stretchr_testify//require",
+    ],
+)
